+++ examples/audio_file_example/android/app/src/main/kotlin/com/example/audio_file_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.audio_file_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}