+++ lib/google_speech.dart
@@ -1,0 +1,5 @@
+export 'exception.dart';
+export 'recognition_config.dart';
+export 'speech_client_authenticator.dart';
+export 'speech_to_text.dart';
+export 'utils.dart';