+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## [0.0.1] - TODO: Add release date.
+## [0.3.0] - TODO: Add SpeechToText Api.
+* A sample project was added.
 
-* TODO: Describe initial release.
+## [0.2.0] - TODO: Add SpeechToText Api.
+* The function to use the Google Speech Api via stream was added.
+
+## [0.1.0] - TODO: Add SpeechToText Api.
+* The function to use the Google Speech Api via request was added.